[DEBT-043-D4] Split cli/alerts.py → cli/alerts/ package
- Split 521-line cli/alerts.py into 7 focused modules:
  - _helpers.py (116 LoC): Storage helpers, daemon spawning
  - list_cmd.py (35 LoC): alerts list command
  - add_cmd.py (76 LoC): alerts add command
  - remove.py (27 LoC): alerts remove command
  - monitor.py (218 LoC): alerts monitor command + async helpers
  - trim_log.py (95 LoC): alerts trim-log command
  - __init__.py (35 LoC): Package exports + Typer app registration
- Updated test_alerts.py mock paths for new package structure
- All files under 400 LoC ceiling (largest: monitor.py at 218 LoC)
- Quality gates: pre-commit, pytest (1052 passed)
