+++ src/kalshi_research/cli/alerts/list_cmd.py
@@ -1,0 +1,35 @@
+"""List alerts command."""
+
+from rich.table import Table
+
+from kalshi_research.cli.alerts._helpers import load_alerts
+from kalshi_research.cli.utils import console
+
+
+def alerts_list() -> None:
+    """List all active alerts."""
+    data = load_alerts()
+    conditions = data.get("conditions", [])
+
+    if not conditions:
+        console.print("[yellow]No active alerts.[/yellow]")
+        return
+
+    table = Table(title="Active Alerts")
+    table.add_column("ID", style="cyan")
+    table.add_column("Type", style="green")
+    table.add_column("Ticker", style="white")
+    table.add_column("Threshold", style="yellow")
+    table.add_column("Label", style="dim")
+
+    for cond in conditions:
+        table.add_row(
+            cond["id"][:8],
+            cond["condition_type"],
+            cond["ticker"],
+            str(cond["threshold"]),
+            cond.get("label", ""),
+        )
+
+    console.print(table)
+    console.print(f"\n[dim]Total: {len(conditions)} alerts[/dim]")
